Remove useGlobalOpenTelemetry option
