--- instrumentation/log4j/log4j-appender-2.17/library/src/test/resources/log4j2-use-global.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Configuration status="WARN"
-  packages="com.example.appender,io.opentelemetry.instrumentation.log4j.appender.v2_17">
-  <Appenders>
-    <Console name="Console" target="SYSTEM_OUT">
-      <PatternLayout
-        pattern="%d{HH:mm:ss.SSS} [%t] %-5level %logger{36} traceId: %X{trace_id} spanId: %X{span_id} flags: %X{trace_flags} - %msg%n"/>
-    </Console>
-    <OpenTelemetry name="OpenTelemetryAppender" useGlobalOpenTelemetry="true" captureMapMessageAttributes="true" captureMarkerAttribute="true" captureContextDataAttributes="*" />
-  </Appenders>
-  <Loggers>
-    <Logger name="TestLogger" level="All">
-      <AppenderRef ref="OpenTelemetryAppender" level="All"/>
-      <AppenderRef ref="Console" level="All"/>
-    </Logger>
-    <Root>
-      <AppenderRef ref="Console" level="All"/>
-    </Root>
-  </Loggers>
-</Configuration>
