+++ instrumentation/logback/logback-appender-1.0/library/README.md
@@ -61,8 +61,8 @@ The following demonstrates how you might configure the appender in your `logback
 In this example Logback log events will be sent to both the console appender and
 the `OpenTelemetryAppender`.
 
-In order to function, `OpenTelemetryAppender` needs access to an `OpenTelemetry` instance. This can
-be programmatically set during application startup as follows:
+In order to function, `OpenTelemetryAppender` needs access to an `OpenTelemetry` instance. This must
+be set programmatically during application startup as follows:
 
 ```java
 import io.opentelemetry.instrumentation.logback.OpenTelemetryAppender;
@@ -80,22 +80,3 @@ public class Application {
   }
 }
 ```
-
-Optionally, you can configure `OpenTelemetryAppender` to use the `GlobalOpenTelemetry` instance:
-```xml
-<appender name="OpenTelemetry"
-  class="io.opentelemetry.instrumentation.logback.appender.v1_0.OpenTelemetryAppender">
-  <useGlobalOpenTelemetry>true</useGlobalOpenTelemetry>
-</appender>
-```
-
-Note: Setting `useGlobalOpenTelemetry` causes `OpenTelemetryAppender` to
-call `GlobalOpenTelemetry.get()`, which initializes `GlobalOpenTelemetry` and prevents future calls
-to `GlobalOpenTelemetry.set(...)`. `GlobalOpenTelemetry.get()` will return a noop instance unless
-you've opted
-into [autoconfigure](https://github.com/open-telemetry/opentelemetry-java/tree/main/sdk-extensions/autoconfigure)
-via:
-
-- Add dependency on `io.opentelemetry:opentelemetry-sdk-extension-autoconfigure:OPENTELEMETRY_VERSION`
-- Opt into `GlobalOpenTelemetry` autoconfigure by
-  setting `-Dotel.java.global-autoconfigure.enabled=true`
