--- instrumentation/logback/logback-appender-1.0/library/src/test/resources/logback-use-global.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-
-  <appender name="console" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>
-        %d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n
-      </pattern>
-    </encoder>
-  </appender>
-  <appender name="OpenTelemetry"
-    class="io.opentelemetry.instrumentation.logback.appender.v1_0.OpenTelemetryAppender">
-    <captureExperimentalAttributes>false</captureExperimentalAttributes>
-    <captureCodeAttributes>true</captureCodeAttributes>
-    <captureMarkerAttribute>true</captureMarkerAttribute>
-    <captureMdcAttributes>*</captureMdcAttributes>
-    <useGlobalOpenTelemetry>true</useGlobalOpenTelemetry>
-  </appender>
-
-  <root level="INFO">
-    <appender-ref ref="console"/>
-    <appender-ref ref="OpenTelemetry"/>
-  </root>
-
-</configuration>
